divided change button to 3 components and their css
