+++ src/app/components/UI/ChangeEmailModal.module.css
@@ -1,0 +1,148 @@
+@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;1,100;1,200;1,300;1,400;1,500;1,600;1,700&display=swap');
+
+  .modaloverlay{
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    background-color: var(--modaloverlaybackground);
+    z-index: 1000; 
+    top: 0px;
+    left: 0px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .Xbutton:hover,.darkbutton:hover{
+    cursor: pointer;
+  }
+
+  .leftflex{
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: center;
+    width: 100%;
+    margin-top: 20px;
+  }
+
+  .modal{
+    background: white;
+    position: relative;
+    padding: 20px;
+    color: var(--mainblackcolor);
+    font-family: 'IBM Plex Sans';
+    border-radius: 8px;
+    text-align: left;
+    width:400px;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: flex-start;
+    z-index: 1001;
+  }
+
+  .inputs:focus{
+    border: 0.5px solid var(--brightcolor);
+  }
+
+  .inputs+label,.invalidInput+label{
+    font-family: 'Noto Sans', sans-serif;
+    color: rgb(124, 124, 124);
+    font-size: 10px;
+    font-weight: 500;
+    line-height: 12px;
+    letter-spacing: 0.5px;
+    padding: 14px;
+    position: absolute;
+    pointer-events: none;
+    transition: 0.3s ease all;
+    top: 4px;
+    left: 5px;
+  }
+  
+  .inputs:focus+label,
+  .inputs:valid+label,
+  .inputs:not(:placeholder-shown) + label,
+  .invalidInput:focus+label,
+  .invalidInput:valid+label,
+  .invalidInput.inputs:not(:placeholder-shown)+label{
+    font-size: 8px;
+    top:-5px;
+    left: 5px;
+  }
+
+  .changeformlabel{
+    margin-left:45px ;
+    margin-top: -40px;
+  }
+  
+  .inputs,.invalidInput{
+    min-height: 40px;
+    width: 360px;
+    font-size: 10px;
+    margin-bottom: 20px;
+    text-indent: 16px;
+    border-radius:5px;
+    border-width:0.5px;
+    border-color:var(--brightcolor)
+  }
+
+  .invalidInput{
+    border: 2px solid red;
+  }
+
+  .errorMessage{
+    margin: 0px;
+    padding: 0px;
+    color: red;
+    font-size: 12px;
+    margin-top: -20px;
+    margin-bottom: 15px;
+  }
+  
+  .darkbutton{
+    position: relative;
+    border-radius: 30px;
+    border: none;
+    background-color: var(--enabledbuttonbackground);
+    color: white;
+    font-size: 14px;
+    font-weight: bold;
+    width: 120px;
+    height: 30px;
+    margin-left: 10px
+  }
+
+  .darkbutton:disabled{
+    background-color: var(--disabledbuttonbackground);
+    color: var(--disabledbuttoncolor);
+  }
+  .darkbutton:disabled:hover{
+    cursor: not-allowed;
+  }
+  
+  
+  .mailimage{
+    width:50px;
+    height: 50px;
+  }
+  
+  .updateemailform{
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    margin-left: 0px;
+    margin-right: 40px;
+    width: 100%;
+    height: 200px;
+  }
+  
+  .Xbutton{
+    position: absolute;
+    top: 16px;
+    right: 16px;
+    background-color: white;
+    border: none;
+    font-size: medium;
+  }
