+++ src/app/components/UI/Changebutton.module.css
@@ -28,216 +28,7 @@
     align-items: center;
     width: 100%
   }
-  .Xbutton:hover,.darkbutton:hover{
-    cursor: pointer;
-  }
-
-  .modaloverlay{
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    background-color: var(--modaloverlaybackground);
-    z-index: 1000; 
-    top: 0px;
-    left: 0px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .leftflex{
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-    width: 100%;
-    margin-top: 20px;
-  }
-
-  .modal,.passwordmodal{
-    background: white;
-    position: relative;
-    padding: 20px;
-    color: var(--mainblackcolor);
-    font-family: 'IBM Plex Sans';
-    border-radius: 8px;
-    text-align: left;
-    width:400px;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    z-index: 1001;
-  }
-
-  .rowflex{
-    display: flex;
-    flex-direction:row;
-    align-items: center;
-  }
-
-  .passwordmodal{
-    background: #fff;
-    border-radius: 4px;
-    box-shadow: 1px 7px 20px 2px rgba(0,0,0,.4);
-    height: 650px;
-    left: 50%;
-    overflow: hidden;
-    pointer-events: all;
-    position: fixed;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    width: 850px;
-    z-index: 111;
-    padding: 0px;
-  }
-
-  .inputwrap{
-    width:100%;
-    position: relative;
-  }
-
-  .inputs:focus{
-    border: 0.5px solid var(--brightcolor);
-  }
-
-  .inputs+label,.invalidInput+label{
-    font-family: 'Noto Sans', sans-serif;
-    color: rgb(124, 124, 124);
-    font-size: 10px;
-    font-weight: 500;
-    line-height: 12px;
-    letter-spacing: 0.5px;
-    padding: 14px;
-    position: absolute;
-    pointer-events: none;
-    transition: 0.3s ease all;
-    top: 4px;
-    left: 5px;
-  }
   
-  .inputs:focus+label,
-  .inputs:valid+label,
-  .inputs:not(:placeholder-shown) + label,
-  .invalidInput:focus+label,
-  .invalidInput:valid+label,
-  .invalidInput.inputs:not(:placeholder-shown)+label{
-    font-size: 8px;
-    top:-5px;
-    left: 5px;
-  }
-
-  .formcontainer{
-    padding: 24px;
-    min-width: 260px;
-    max-width: 440px;
-    max-height: 600px;
-  }
-
-  .changeformlabel{
-    margin-left:45px ;
-    margin-top: -40px;
-  }
-  
-  .inputs,.invalidInput{
-    min-height: 40px;
-    width: 360px;
-    font-size: 10px;
-    margin-bottom: 20px;
-  }
-
-  .invalidInput{
-    border: 2px solid red;
-  }
-
-
-  .errorMessage{
-    margin: 0px;
-    padding: 0px;
-    color: red;
-    font-size: 12px;
-    margin-top: -20px;
-    margin-bottom: 15px;
-  }
-  
-  .darkbutton{
-    position: relative;
-    border-radius: 30px;
-    border: none;
-    background-color: var(--enabledbuttonbackground);
-    color: white;
-    font-size: 14px;
-    font-weight: bold;
-    width: 120px;
-    height: 30px;
-    margin-left: 10px
-  }
-
-  .savebutton{
-    font-family: 'Noto Sans', sans-serif;
-    font-size: 14px;
-    border: none;
-    font-weight: 700;
-    letter-spacing: unset;
-    line-height: 18px;
-    text-transform: unset;
-    max-width: 392px;
-    width: auto;
-    min-width: 155px;
-    background: var(--enabledbuttonbackground);
-    border-radius: 999px;
-    color: white;
-    height: 40px;
-    padding: 0 16px;
-  }
-
-  .savebutton:hover{
-    cursor: pointer;
-  }
-  
-  
-  .darkbutton:disabled{
-    background-color: var(--disabledbuttonbackground);
-    color: var(--disabledbuttoncolor);
-  }
-  .darkbutton:disabled:hover{
-    cursor: not-allowed;
-  }
-  
-  
-  .mailimage{
-    width:50px;
-    height: 50px;
-  }
-  
-  .updateemailform,.changepasswordform{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    margin-left: 0px;
-    margin-right: 40px;
-    width: 100%;
-    height: 200px;
-  }
-  
-  .changepasswordform{
-    height: 110px;
-  }
-  
-  
-  .inlineflex{
-    display: inline-block;
-  }
-  
-  .Xbutton{
-    position: absolute;
-    top: 16px;
-    right: 16px;
-    background-color: white;
-    border: none;
-    font-size: medium;
-  }
-
   .brightbutton{
     width: 80px;
     background-color: var(--buttonbackground);
